Calculate Kroll's Linearity Test

Calculate a statistical test for linearity from Kroll et. al. (2000) doi:10.5858/2000-124-1331-EOTEON using average deviation from linearity

Usage

calculate_adl_kroll_test(dilution_data, conc_var, signal_var)

Arguments

dilution_data

A data frame or tibble containing dilution data

conc_var

Column name in dilution_data to indicate concentration

signal_var

Column name in dilution_data to indicate signal

Value

A data frame of statistical results from Kroll et. al. (2000) doi:10.5858/2000-124-1331-EOTEON

adl_kroll: Average deviation from linearity as defined in Kroll et. al. (2000) doi:10.5858/2000-124-1331-EOTEON

precision_on_percent_scale: Estimate of precision on percent scale. It is the ratio of the estimated regression standard error from the best-fit polynomial and the estimated solution mean.

uncorrected_critical_value: Critical value (Uncorrected) as defined in Kroll et. al. (2000) doi:10.5858/2000-124-1331-EOTEON

corrected_critical_value: Critical value (Corrected) as defined in Hsieh et. al. (2008) doi:10.1080/10543400802071378

uncorrected_kroll_results: Indicate if adl_kroll is less than uncorrected_critical_value If FALSE, there is sufficient evidence to reject the hypothesis that the points are linear.

corrected_kroll_results: Indicate if adl_kroll is less than corrected_critical_value If FALSE, there is sufficient evidence to reject the hypothesis that the points are linear.

best_model Gives values linear, quadratic or cubic to indicate the model that "best fit" the points. The function compare_performance is used to determine the best fit.

Details

The function will return NA if the number of dilution points is less than or equal to three.

References

Kroll, M. H., Praestgaard, J., Michaliszyn, E., & Styer, P. E. (2000). Evaluation of the extent of nonlinearity in reportable range studies. Archives of pathology & laboratory medicine, 124(9), 1331–1338. doi:10.5858/2000-124-1331-EOTEON

Hsieh E., & Liu J. P. (2008). On Statistical Evaluation of the Linearity in Assay Validation Journal of Biopharmaceutical Statistics, 18(4), 677–690. doi:10.1080/10543400802071378

Examples

# Data from Kroll's 2000 paper
solution_number <- c(
  1, 1, 2, 2, 3, 3, 4, 4,
  5, 5, 6, 6, 7, 7
)
result <- c(
  352, 348, 1009, 991, 1603, 1584, 3100, 3200,
  4482, 4390, 5101, 5046, 5669, 5516
)

dilution_data <- data.frame(
  Solution_Number = solution_number,
  Result = result
)

adl_result <- calculate_adl_kroll_test(
  dilution_data,
  "Solution_Number",
  "Result"
)
adl_result
#> # A tibble: 1 × 7
#>   adl_kroll precision_on_percent_scale uncorre…¹ corre…² uncor…³ corre…⁴ best_…⁵
#>       <dbl>                      <dbl>     <dbl>   <dbl> <chr>   <chr>   <chr>  
#> 1      8.63                       5.54      7.61    2.86 FALSE   FALSE   cubic  
#> # … with abbreviated variable names ¹​uncorrected_critical_value,
#> #   ²​corrected_critical_value, ³​uncorrected_kroll_results,
#> #   ⁴​corrected_kroll_results, ⁵​best_model