Create Dilution Table

Create Dilution Table from two data frame or tibble inputs dilution_annot and lipid_data_wide. Both input must have at least one common column. The default common column names is Sample_Name.

Usage

create_dilution_table(
  dilution_annot,
  lipid_data_wide,
  common_column = c("Sample_Name"),
  signal_var = "Area",
  column_group = "Transition_Name"
)

Arguments

dilution_annot

A data frame or tibble that contains information of your dilution batch. A column with sample name should be present.

lipid_data_wide

A wide format data frame or tibble that contains the Sample Name usually at the first column followed by different Transitions/Lipids. Each Transition/Lipid is meant to provide a signal_var for each sample.

common_column

A vector consisting of common column names that must be found in both dilution_annot and lipid_data_wide, Default: c("Sample_Name")

signal_var

Value provided in lipid_data_wide for each transition or lipid which will be used as a column name when the merged data is converted to a long format, Default: 'Area'

column_group

Value used to group the transition column names in lipid_data_wide when the merged data is converted to a long format, Default: 'Transition_Name'

Value

A data frame or tibble in long format

Details

We first merge the dilution_annot with the lipid_data_wide via one or more common columns. Next we convert the data from a wide to a long format. Merging with a Sample Annotation data can be done after this

Examples

# Data Creation
dilution_percent <- c(
  10, 20, 40, 60, 80, 100,
  10, 20, 40, 60, 80, 100
)
dilution_batch_name <- c(
  "B1", "B1", "B1", "B1", "B1", "B1",
  "B2", "B2", "B2", "B2", "B2", "B2"
)
sample_name <- c(
  "Sample_010a", "Sample_020a", "Sample_040a",
  "Sample_060a", "Sample_080a", "Sample_100a",
  "Sample_010b", "Sample_020b", "Sample_040b",
  "Sample_060b", "Sample_080b", "Sample_100b"
)
lipid1_area <- c(
  22561, 31178, 39981, 48390, 52171, 53410,
  32561, 41178, 49981, 58390, 62171, 63410
)
lipid2_area <- c(
  2299075, 4136350, 7020062, 8922063, 9288742, 11365710,
  2300075, 4137350, 7021062, 8923063, 9289742, 11366710
)
dilution_annot <- tibble::tibble(
  Sample_Name = sample_name,
  Dilution_Batch_Name = dilution_batch_name,
  Dilution_Percent = dilution_percent
)
lipid_data <- tibble::tibble(
  Sample_Name = sample_name,
  Lipid1 = lipid1_area,
  Lipid2 = lipid2_area
)
# Create dilution table
dilution_table <- create_dilution_table(dilution_annot, lipid_data,
  common_column = "Sample_Name",
  signal_var = "Area",
  column_group = "Transition_Name"
)
print(dilution_table, width = 100)
#> # A tibble: 24 × 5
#>    Sample_Name Dilution_Batch_Name Dilution_Percent Transition_Name    Area
#>    <chr>       <chr>                          <dbl> <chr>             <dbl>
#>  1 Sample_010a B1                                10 Lipid1            22561
#>  2 Sample_010a B1                                10 Lipid2          2299075
#>  3 Sample_020a B1                                20 Lipid1            31178
#>  4 Sample_020a B1                                20 Lipid2          4136350
#>  5 Sample_040a B1                                40 Lipid1            39981
#>  6 Sample_040a B1                                40 Lipid2          7020062
#>  7 Sample_060a B1                                60 Lipid1            48390
#>  8 Sample_060a B1                                60 Lipid2          8922063
#>  9 Sample_080a B1                                80 Lipid1            52171
#> 10 Sample_080a B1                                80 Lipid2          9288742
#> # … with 14 more rows