Summarize Dilution Table

Get the summary statistics of each group from the dilution table

summarise_dilution_table(
  dilution_table,
  grouping_variable = c("Transition_Name", "Dilution_Batch_Name"),
  conc_var = "Dilution_Percent",
  signal_var = "Area"
)

Arguments

dilution_table

Output given from the function create_dilution_table It is in long table format with columns indicating at least the lipid/transition name, the concentration and signal. Other columns may be present if it is used to group the dilution curve together

grouping_variable

A character vector of column names in dilution_tableto indicate how each dilution curve should be grouped by.

conc_var

Column name in dilution_table to indicate concentration Default: 'Dilution_Percent'

signal_var

Column name in dilution_table to indicate signal Default: 'Area'

Value

A dilution summary table output from the function summarise_dilution_data() for each group

Details

The function first perform tidyr::nest on dilution_table based on the grouping_variable to organise the dilution curve data for each group. Next for each group, the function summarise_dilution_data() is used to get the summary statistics.

Examples

# Data Creation
dilution_percent <- c(
  10, 20, 40, 60, 80, 100,
  10, 20, 40, 60, 80, 100
)
dilution_batch_name <- c(
  "B1", "B1", "B1", "B1", "B1", "B1",
  "B2", "B2", "B2", "B2", "B2", "B2"
)
sample_name <- c(
  "Sample_010a", "Sample_020a", "Sample_040a",
  "Sample_060a", "Sample_080a", "Sample_100a",
  "Sample_010b", "Sample_020b", "Sample_040b",
  "Sample_060b", "Sample_080b", "Sample_100b"
)
lipid1_area <- c(
  22561, 31178, 39981, 48390, 52171, 53410,
  32561, 41178, 49981, 58390, 62171, 63410
)
lipid2_area <- c(
  2299075, 4136350, 7020062, 8922063, 9288742, 11365710,
  2300075, 4137350, 7021062, 8923063, 9289742, 11366710
)
dilution_annot <- tibble::tibble(
  Sample_Name = sample_name,
  Dilution_Batch_Name = dilution_batch_name,
  Dilution_Percent = dilution_percent
)
lipid_data <- tibble::tibble(
  Sample_Name = sample_name,
  Lipid1 = lipid1_area,
  Lipid2 = lipid2_area
)
# Create dilution table
dilution_table <- create_dilution_table(dilution_annot, lipid_data,
  common_column = "Sample_Name",
  signal_var = "Area",
  column_group = "Transition_Name"
)
# Give summary result for each dilution curve grouped by Transition_Name
# and Dilution Batch
dilution_summary <- dilution_table %>%
  summarise_dilution_table(
    grouping_variable = c(
      "Transition_Name",
      "Dilution_Batch_Name"
    ),
    conc_var = "Dilution_Percent",
    signal_var = "Area"
  )
print(dilution_summary, width = 100)
#> # A tibble: 4 × 9
#>   Transition_Name Dilution_Batch_Name r_corr r2_linear r2_adj_linear
#>   <chr>           <chr>                <dbl>     <dbl>         <dbl>
#> 1 Lipid1          B1                   0.956     0.913         0.892
#> 2 Lipid2          B1                   0.972     0.946         0.932
#> 3 Lipid1          B2                   0.956     0.913         0.892
#> 4 Lipid2          B2                   0.972     0.946         0.932
#>   mandel_stats mandel_p_val pra_linear concavity
#>          <dbl>        <dbl>      <dbl>     <dbl>
#> 1        70.1       0.00357       66.4     -4.09
#> 2         5.67      0.0974        74.8   -735.  
#> 3        70.1       0.00357       66.4     -4.09
#> 4         5.67      0.0974        74.8   -735.