Update Cognostics Manually

Update cognostics on dilution_summary based on the cognostics parameters given by cog_df. We assume cog_df is created manually

update_cog_manual(
  dilution_summary,
  cog_df,
  col_name_vec = "col_name_vec",
  desc_vec = "desc_vec",
  type_vec = "type_vec"
)

Arguments

dilution_summary

The summary data frame or tibble generated by function summarise_dilution_table() and/or evaluate_linearity().

cog_df

A data frame or tibble output from the function create_default_cog_df() or created manually

col_name_vec

Column name in cog_df to indicate the columns in dilution_summary that needs to be converted to a cognostics, Default: 'col_name_vec'

desc_vec

Column name in cog_df to indicate the description for each cognostics as define in trelliscopejs::cog, Default: 'desc_vec'

type_vec

Column name in cog_df to indicate the type of each cognostics as define in trelliscopejs::cog, Default: 'type_vec'

Value

dilution_summary with some columns converted to type cog as defined in trelliscopejs::cog

Details

More details in trelliscopejs::cog can be found in https://rdrr.io/cran/trelliscopejs/man/cog.html

Examples

dilution_percent <- c(
  10, 20, 25, 40, 50, 60,
  75, 80, 100, 125, 150,
  10, 25, 40, 50, 60,
  75, 80, 100, 125, 150
)
dilution_batch_name <- c(
  "B1", "B1", "B1", "B1", "B1",
  "B1", "B1", "B1", "B1", "B1", "B1",
  "B2", "B2", "B2", "B2", "B2",
  "B2", "B2", "B2", "B2", "B2"
)
sample_name <- c(
  "Sample_010a", "Sample_020a",
  "Sample_025a", "Sample_040a", "Sample_050a",
  "Sample_060a", "Sample_075a", "Sample_080a",
  "Sample_100a", "Sample_125a", "Sample_150a",
  "Sample_010b", "Sample_025b",
  "Sample_040b", "Sample_050b", "Sample_060b",
  "Sample_075b", "Sample_080b", "Sample_100b",
  "Sample_125b", "Sample_150b"
)
lipid1_area_saturated <- c(
  5748124, 16616414, 21702718, 36191617,
  49324541, 55618266, 66947588, 74964771,
  75438063, 91770737, 94692060,
  5192648, 16594991, 32507833, 46499896,
  55388856, 62505210, 62778078, 72158161,
  78044338, 86158414
)
lipid2_area_linear <- c(
  31538, 53709, 69990, 101977, 146436, 180960,
  232881, 283780, 298289, 344519, 430432,
  25463, 63387, 90624, 131274, 138069,
  205353, 202407, 260205, 292257, 367924
)
lipid3_area_lod <- c(
  544, 397, 829, 1437, 1808, 2231,
  3343, 2915, 5268, 8031, 11045,
  500, 903, 1267, 2031, 2100,
  3563, 4500, 5300, 8500, 10430
)
lipid4_area_nonlinear <- c(
  380519, 485372, 478770, 474467, 531640, 576301,
  501068, 550201, 515110, 499543, 474745,
  197417, 322846, 478398, 423174, 418577,
  426089, 413292, 450190, 415309, 457618
)

dilution_annot <- tibble::tibble(
  Sample_Name = sample_name,
  Dilution_Batch_Name = dilution_batch_name,
  Dilution_Percent = dilution_percent
)
lipid_data <- tibble::tibble(
  Sample_Name = sample_name,
  Lipid1 = lipid1_area_saturated,
  Lipid2 = lipid2_area_linear,
  Lipid3 = lipid3_area_lod,
  Lipid4 = lipid4_area_nonlinear
)

# Create dilution table
dilution_table <- create_dilution_table(dilution_annot, lipid_data,
  common_column = "Sample_Name",
  signal_var = "Area",
  column_group = "Transition_Name"
)

# Create dilution table and dilution statistical summary
dilution_summary <- dilution_table %>%
  summarise_dilution_table(
    grouping_variable = c(
      "Transition_Name",
      "Dilution_Batch_Name"
    ),
    conc_var = "Dilution_Percent",
    signal_var = "Area"
  ) %>%
  dplyr::arrange(.data$Transition_Name) %>%
  evaluate_linearity(grouping_variable = c(
    "Transition_Name",
    "Dilution_Batch_Name"
  ))

# Create our own cog_df
col_name_vec <- c(
  "Transition_Name", "Dilution_Batch_Name",
  "Transition_Name_Class"
)

desc_vec <- c(
  "Transition_Name", "Dilution_Batch_Name",
  "Transition_Name_Class"
)

type_vec <- c("factor", "factor", "factor")

cog_df <- data.frame(
  col_name_vec = col_name_vec,
  desc_vec = desc_vec,
  type_vec = type_vec
)

updated_summary <- update_cog_manual(dilution_summary, cog_df,
  col_name_vec = "col_name_vec",
  desc_vec = "desc_vec",
  type_vec = "type_vec"
)

# Observe that the first two columns has been converted
# to class cognostics
dilution_summary
#> # A tibble: 8 × 11
#>   Transition_N…¹ Dilut…² wf1_g…³ wf2_g…⁴ r_corr pra_l…⁵ mande…⁶ concav…⁷ r2_li…⁸
#>   <chr>          <chr>   <chr>   <chr>    <dbl>   <dbl>   <dbl>    <dbl>   <dbl>
#> 1 Lipid1         B1      Poor L… Satura…  0.963    70.5 2.97e-5 -4.17e+3  0.928 
#> 2 Lipid1         B2      Poor L… Satura…  0.950    62.3 1.66e-4 -4.14e+3  0.903 
#> 3 Lipid2         B1      Good L… Good L…  0.990    92.8 1.50e-1 -4.91e+0  0.980 
#> 4 Lipid2         B2      Good L… Good L…  0.995    94.3 3.82e-1 -1.94e+0  0.990 
#> 5 Lipid3         B1      Poor L… LOD      0.964    71.2 6.78e-6  4.68e-1  0.930 
#> 6 Lipid3         B2      Poor L… LOD      0.978    74.7 2.56e-3  3.21e-1  0.956 
#> 7 Lipid4         B1      Poor L… Poor L…  0.311  -251.  6.60e-3 -2.05e+1  0.0970
#> 8 Lipid4         B2      Poor L… Poor L…  0.608   -73.1 5.33e-2 -2.29e+1  0.370 
#> # … with 2 more variables: r2_adj_linear <dbl>, mandel_stats <dbl>, and
#> #   abbreviated variable names ¹​Transition_Name, ²​Dilution_Batch_Name,
#> #   ³​wf1_group, ⁴​wf2_group, ⁵​pra_linear, ⁶​mandel_p_val, ⁷​concavity, ⁸​r2_linear
updated_summary
#> # A tibble: 8 × 11
#>   Transition_N…¹ Dilut…² wf1_g…³ wf2_g…⁴ r_corr pra_l…⁵ mande…⁶ concav…⁷ r2_li…⁸
#>   <cog>          <cog>   <chr>   <chr>    <dbl>   <dbl>   <dbl>    <dbl>   <dbl>
#> 1 Lipid1         B1      Poor L… Satura…  0.963    70.5 2.97e-5 -4.17e+3  0.928 
#> 2 Lipid1         B2      Poor L… Satura…  0.950    62.3 1.66e-4 -4.14e+3  0.903 
#> 3 Lipid2         B1      Good L… Good L…  0.990    92.8 1.50e-1 -4.91e+0  0.980 
#> 4 Lipid2         B2      Good L… Good L…  0.995    94.3 3.82e-1 -1.94e+0  0.990 
#> 5 Lipid3         B1      Poor L… LOD      0.964    71.2 6.78e-6  4.68e-1  0.930 
#> 6 Lipid3         B2      Poor L… LOD      0.978    74.7 2.56e-3  3.21e-1  0.956 
#> 7 Lipid4         B1      Poor L… Poor L…  0.311  -251.  6.60e-3 -2.05e+1  0.0970
#> 8 Lipid4         B2      Poor L… Poor L…  0.608   -73.1 5.33e-2 -2.29e+1  0.370 
#> # … with 2 more variables: r2_adj_linear <dbl>, mandel_stats <dbl>, and
#> #   abbreviated variable names ¹​Transition_Name, ²​Dilution_Batch_Name,
#> #   ³​wf1_group, ⁴​wf2_group, ⁵​pra_linear, ⁶​mandel_p_val, ⁷​concavity, ⁸​r2_linear