Validate Curve Data Wide

Validate Curve Data in wide form

Usage

validate_curve_data_wide(curve_data_wide, needed_column = c("Sample_Name"))

Arguments

curve_data_wide

A wide format data frame or tibble that contains the Sample Name usually at the first column followed by different curves. Each curve is meant to provide a signal_var for each sample.

needed_column

A vector consisting of needed column names that must be found in curve_data_wide, Default: c("Sample_Name")

Value

An error if the things in needed_column

is not found in the Curve Data in wide form.

Examples

sample_name <- c(
  "Sample_010a", "Sample_020a", "Sample_040a",
  "Sample_060a", "Sample_080a", "Sample_100a",
  "Sample_010b", "Sample_020b", "Sample_040b",
  "Sample_060b", "Sample_080b", "Sample_100b"
)

curve_1_good_linearity <- c(
  22561, 31178, 39981, 48390, 52171, 53410,
  32561, 41178, 49981, 58390, 62171, 63410
)

curve_2_good_linearity <- c(
  2299075, 4136350, 7020062, 8922063, 9288742, 11365710,
  2300075, 4137350, 7021062, 8923063, 9289742, 11366710
)

curve_data_wide <- tibble::tibble(
  Sample_Name = sample_name,
  `Curve_1` = curve_1_good_linearity,
  `Curve_2` = curve_2_good_linearity
)

validate_curve_data_wide(
  curve_data_wide = curve_data_wide,
  needed_column = c("Sample_Name")
)